% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_table.R
\name{data_statistic}
\alias{data_statistic}
\title{This function draw table}
\usage{
data_statistic(
  df,
  gvar = NULL,
  varlist = NULL,
  skewvar = NULL,
  cat_var = NULL,
  out_docx,
  path,
  pnormtest = 0.05
)
}
\arguments{
\item{df}{data, gvar:group,varlist:variate,skewvar:Skew variable,cat_var:cat_var,out_docx:filename.}
}
\value{
table.
}
\description{
This function draw table
}
