% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eaf_from_1000G.R
\name{get_eaf_from_1000G}
\alias{get_eaf_from_1000G}
\title{My Function}
\usage{
get_eaf_from_1000G(dat, path, type = "exposure")
}
\arguments{
\item{dat}{mr dat, path file,.}
}
\value{
data.frame.
}
\description{
This function matching eaf.
}
