% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_cal.R
\name{med_cal}
\alias{med_cal}
\title{My Function}
\usage{
med_cal(beta1, se1, beta2, se2, beta_total, n = 1000)
}
\arguments{
\item{beta1, }{se1 beta2,se2,beta_total,n.}
}
\value{
data.frame.
}
\description{
This function calculate mediation effection.
}
